Combine Tag Manager settings edit stories from SettingsForm to SettingsEdit.
